requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.7
openai>=1.40.0
chromadb>=0.5.4
tiktoken>=0.7.0
python-dotenv>=1.0.1
langchain-chroma
rank_bm25
streamlit